Add start/create chat functionality
The create chat button opens a modal that contains the user contacts.
When a contact is clicked the chatroom is opened for the conversation to
start.
Also ensured that chats are updated after sending a message to ensure that
the chats are effectively updated when a new message is sent.